requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=13.0
eth-account>=0.10.0
solders>=0.20.0
aiohttp>=3.9.0
psycopg[binary,pool]>=3.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
